backend/main/requirements.txt
Flask==2.3.3
PyMySQL==1.1.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
PyYAML==6.0.1
cozepy==0.7.0
DBUtils==3.1.2
gunicorn==21.2.0
pandas==2.1.4
openpyxl==3.1.2
python-keycloak==5.8.1
flask-sock==0.7.0
simple-websocket==1.0.0
redis==5.0.1
requests==2.31.0
Werkzeug==2.3.7
